`Updated Browsers.py and Discord.py with Roblox API changes`
Updated `Browsers.py` and `Discord.py` to reflect changes in the handling of Roblox user information. In `Browsers.py`, imported the `requests` module and added the `robloxinfo(self, webhook)` method to the `Browsers` class. This new method checks the Roblox configuration, retrieves user information from the Roblox API, and sends a POST request to the specified webhook. Removed the `robloxinfo(self, webhook)` method and its call from the `Discord` class in `Discord.py`.
